wip removed payload class, use dicts for everything, move around tests
